requirements.txt
chromadb==1.0.20
langgraph==0.6.6
pydantic==2.11.7
# PyYAML==6.0.1
PyYAML==6.0.2
rich==14.1.0
sentence_transformers==3.0.1
